backend/requirements.txt
# ───────── Web framework ─────────
fastapi==0.110.1
uvicorn==0.25.0
starlette==0.37.2

# ───────── Environment ─────────
python-dotenv==1.2.1

# ───────── Auth & security ─────────
bcrypt==4.1.3
PyJWT==2.10.1
python-multipart==0.0.21
python-jose==3.5.0
email-validator==2.3.0

# ───────── Validation ─────────
pydantic==2.12.5

# ───────── Database ─────────
aiomysql
pymysql
sqlalchemy

# ───────── AI (primary) ─────────
openai==1.99.9
tiktoken==0.12.0

# ───────── Async HTTP ─────────
httpx==0.28.1
anyio==4.12.0

# ───────── Utilities ─────────
rich==14.2.0
